Move Congratulation component to course material
